@@ -29,17 +29,13 @@ setuptools.setup(
         'License :: OSI Approved :: MIT License',
 
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-        'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
-        'Programming Language :: Python :: 3 :: Only',
+        'Programming Language :: Python :: 3.10 :: Only',
     ],
     keywords='smart home, machine learning, house automation',
     packages=packages,
     include_package_data=True,
-    python_requires='>=3.6, <4',
+    python_requires='>=3.10, <4',
     install_requires=[req for req in requirements if req[:2] != "# "],
     entry_points={
         'console_scripts': [
