@@ -3,7 +3,6 @@ from fastapi import APIRouter, Body, Path, status, Depends
 
 from learninghouse.api.errors.brain import BrainExists, BrainNoConfiguration
 from learninghouse.api.errors.sensor import NoSensor, SensorExists
-from learninghouse.core.settings import service_settings
 from learninghouse.models.configuration import (BrainConfiguration,
                                                 BrainConfigurationRequest,
                                                 BrainConfigurations,
@@ -13,7 +12,7 @@ from learninghouse.models.configuration import (BrainConfiguration,
                                                 SensorType)
 from learninghouse.services.configuration import (BrainConfigurationService,
                                                   SensorConfigurationService)
-from learninghouse.services.authorization import protect_admin
+from learninghouse.services.auth import auth_service
 
 brain_router = APIRouter(
     prefix='/brain',
@@ -164,9 +163,8 @@ async def sensor_delete(name: str):
 
 router = APIRouter(
     tags=['configuration'],
-    dependencies=[Depends(protect_admin)]
+    dependencies=[Depends(auth_service().protect_admin)]
 )
 
-if service_settings().api_key_admin:
-    router.include_router(brain_router)
-    router.include_router(sensor_router)
+router.include_router(brain_router)
+router.include_router(sensor_router)
